+++ packages/cli/src/commands/create.ts
@@ -1,19 +1,19 @@
-import { buildProject } from '@/src/utils/build-project';
-import { copyTemplate as copyTemplateUtil } from '@/src/utils/copy-template';
-import { checkServer, handleError } from '@/src/utils/handle-error';
-import { runBunCommand } from '@/src/utils/run-bun';
-import { logger } from '@elizaos/core';
+import {
+  buildProject,
+  copyTemplate as copyTemplateUtil,
+  getElizaDirectories,
+  handleError,
+  promptAndStorePostgresUrl,
+  runBunCommand,
+  setupPgLite,
+} from '@/src/utils';
 import { Command } from 'commander';
-import { execa } from 'execa';
 import { existsSync, readFileSync } from 'node:fs';
 import fs from 'node:fs/promises';
-import os from 'node:os';
 import path from 'node:path';
 import prompts from 'prompts';
 import colors from 'yoctocolors';
 import { z } from 'zod';
-import { displayBanner } from '../utils/displayBanner';
-import { setupPgLite, promptAndStorePostgresUrl, getElizaDirectories } from '../utils/get-config';
 
 /**
  * This module handles creating both projects and plugins.
