+++ packages/cli/src/commands/publish.ts
@@ -1,19 +1,21 @@
-import { promises as fs, existsSync } from 'node:fs';
-import path from 'node:path';
-import { fileURLToPath } from 'node:url';
-import { getGitHubCredentials } from '@/src/utils/github';
-import { handleError } from '@/src/utils/handle-error';
-import { publishToGitHub, testPublishToGitHub, testPublishToNpm } from '@/src/utils/publisher';
+import {
+  getGitHubCredentials,
+  handleError,
+  publishToGitHub,
+  testPublishToGitHub,
+  testPublishToNpm,
+} from '@/src/utils';
 import {
   getRegistrySettings,
   initializeDataDir,
   saveRegistrySettings,
   validateDataDir,
 } from '@/src/utils/registry/index';
-import { logger } from '@elizaos/core';
-import { Octokit } from '@octokit/rest';
 import { Command } from 'commander';
 import { execa } from 'execa';
+import { existsSync, promises as fs } from 'node:fs';
+import path from 'node:path';
+import { fileURLToPath } from 'node:url';
 import prompts from 'prompts';
 // Import performCliUpdate directly for updating CLI
 import { performCliUpdate } from './update-cli';
