+++ packages/cli/src/index.ts
@@ -17,8 +17,8 @@ import { start } from './commands/start';
 import { teeCommand as tee } from './commands/tee';
 import { test } from './commands/test';
 import { update } from './commands/update';
-import { loadEnvironment } from './utils/get-config';
-import { displayBanner, getVersion } from './utils/displayBanner';
+import { loadEnvironment } from './utils';
+import { displayBanner } from './utils/display-banner';
 import { setupMonorepo } from './commands/install';
 import { updateCLI } from './commands/update-cli';
 import { plugins } from './commands/plugins';
