+++ packages/cli/src/utils/index.ts
@@ -1,1 +1,23 @@
+// Core utilities
+export * from './audioBuffer';
+export * from './build-project';
+export * from './cli-prompts';
+export * from './config-manager';
+export * from './copy-template';
+export * from './display-banner';
+export * from './env-prompt';
+export * from './get-package-info';
+export * from './github';
+export * from './handle-error';
+export * from './helpers';
+export * from './install-plugin';
+export * from './load-plugin';
 export * from './monorepo-utils';
+export * from './package-manager';
+export * from './port-handling';
+export * from './publisher';
+export * from './resolve-import';
+export * from './run-bun';
+export * from './testRunner';
+export * from './user-environment';
+export * from './get-config';
